+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 3.0.0 (unreleased)
 
+- Changed `capture_query_stats` to only reset stats for current database in Postgres 12+
+- Changed `reset_query_stats` to only reset stats for current database (use `reset_instance_query_stats` to reset stats for entire instance)
 - Dropped support for Ruby < 2.7 and Rails < 6
 - Dropped support for pg_query < 2